pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tmviz"
version = "0.1.0"
description = "Turing machine simulator and visualizer"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pygame-ce>=2.5.0,<3",
    "transitions>=0.9,<1",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.9,<2",
    "pytest>=8.0,<9",
    "pytest-cov>=5.0,<6",
    "ruff>=0.3,<1",
]

[project.scripts]
tmviz = "tmviz.main:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

